+++ CHANGELOG.rst
@@ -14,6 +14,13 @@ Change Log
 Unreleased
 --------------------
 
+[3.4.30] 2020-07-29
+-------------------
+
+* Add flag to sync updates in an EnterpriseCatalogQuery with its associated EnterpriseCustomerCatalogs.
+* Create a post_save signal to overwrite the content_filter with the update.
+* Changes should also be sent to the Enterprise Catalog service.
+
 [3.4.29] 2020-07-29
 -------------------